UI skeleton
A functional skeleton of the UI. Will make it prettier once I get the assets to do so. Works fine in the preview, but breaks when built due to how Unity handles resources. Will look into a solution to this, but I think it can be fixed by cloning the Resources folder into the final build folder if no better solution appears.